Add different source to compare for objective
The objective can be a fixed amount, a category or a tag. To save the type used, the compare_with column is added (e.g. the value can be fixed, catg or tag). If category or tag is selected, the amount from the previous period for the category or tag is used.